chromium/third_party/blink/web_tests/wpt_internal/display-lock/css-subtree-visibility/subtree-visibility-033.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Subtree Visibility: accessibility</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<meta name="assert" content="subtree-visibility hidden subtree is not exposed to accessibility">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.hidden {
  subtree-visibility: hidden;
}
</style>
<div id="container">
  <div id="target" class=hidden>
    target
    <div id="child">
      child
    </div>
  </div>
</div>

<script>
function axElementById(id) {
  return accessibilityController.accessibleElementById(id);
}

async_test(async (t) => {
  let target = document.getElementById("target");
  let axTarget = axElementById("target");
  t.step(() => { assert_false(axTarget.isIgnored); });
  t.step(() => { assert_equals(axTarget.childrenCount, 0); });

  target.classList.remove("hidden");
  requestAnimationFrame(() => requestAnimationFrame(() => {
    // The ax object for #target got replaced since the layout object changed, so use the new ax object.
    axTarget = axElementById("target");
    t.step(() => { assert_equals(axTarget.childrenCount, 2, "After commit, nodes in target subtree are not ignored"); });
    t.done();
  }));
}, "Nodes in target non-activatable tree are not exposed to accessibility tree");
</script>
